refactor: consolidate from 4 packages to 2
Merged @layer-ai/types and @layer-ai/config into @layer-ai/sdk. Updated all imports across CLI and API.
