+++ apps/api/src/lib/db/redis.ts
@@ -1,5 +1,5 @@
 import Redis from 'ioredis';
-import type { Gate } from '@layer-ai/types';
+import type { Gate } from '@layer-ai/sdk';
 
 // Create redis client
 const redis = new Redis(process.env.REDIS_URL || 'redis://localhost:6379', {
