+++ apps/api/src/services/providers/anthropic.ts
@@ -1,6 +1,6 @@
 import Anthropic from '@anthropic-ai/sdk/index';
-import type { Message, SupportedModel } from '@layer-ai/types';
-import { MODEL_REGISTRY } from '@layer-ai/types';
+import type { Message, SupportedModel } from '@layer-ai/sdk';
+import { MODEL_REGISTRY } from '@layer-ai/sdk';
 
 // Lazy-initialize Anthropic client
 let anthropic: Anthropic | null = null;
