+++ apps/api/src/services/providers/openai.ts
@@ -1,6 +1,6 @@
 import OpenAI from 'openai';
-import type { Message, SupportedModel } from '@layer-ai/types';
-import { MODEL_REGISTRY } from '@layer-ai/types';
+import type { Message, SupportedModel } from '@layer-ai/sdk';
+import { MODEL_REGISTRY } from '@layer-ai/sdk';
 
 // Lazy-initialize OpenAI client
 let openai: OpenAI | null = null;
