+++ packages/cli/package.json
@@ -47,8 +47,6 @@
   },
   "homepage": "https://github.com/micah-nettey/layer-ai#readme",
   "dependencies": {
-    "@layer-ai/types": "workspace:*",
-    "@layer-ai/config": "workspace:*",
     "@layer-ai/sdk": "workspace:*",
     "commander": "^11.1.0",
     "chalk": "^5.3.0",
