+++ packages/sdk/package.json
@@ -9,6 +9,10 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./config": {
+      "types": "./dist/config/index.d.ts",
+      "import": "./dist/config/index.js"
     }
   },
   "files": [
@@ -47,9 +51,11 @@
   },
   "homepage": "https://github.com/micah-nettey/layer-ai#readme",
   "dependencies": {
-    "@layer-ai/types": "workspace:*"
+    "yaml": "^2.3.4",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@types/node": "^20.10.0",
     "typescript": "^5.9.3"
   }
 }
