+++ packages/sdk/src/config/index.ts
@@ -2,5 +2,5 @@ export { loadConfig, parseConfig } from './loader.js';
 export { validateConfig } from './validator.js';
 export { parseYAML } from './parser.js';
 
-// Re-export types from @layer-ai/types for convenience
-export type { LayerConfigFile, GateConfig } from '@layer-ai/types';
+// Re-export types from SDK types for convenience
+export type { LayerConfigFile, GateConfig } from '../types/index.js';
