+++ packages/sdk/src/index.ts
@@ -3,18 +3,6 @@ export type { LayerConfig } from './types.js';
 export { KeysResource } from './resources/keys.js';
 export { GatesResource } from './resources/gates.js';
 export { LogsResource } from './resources/logs.js';
-export type { 
-  Message, 
-  CompletionRequest, 
-  CompletionResponse, 
-  ErrorResponse,
-  Gate,
-  CreateGateRequest,
-  UpdateGateRequest,
-  ApiKey,
-  CreateKeyRequest,
-  CreateKeyResponse,
-  Log,
-  ListLogOptions,
-  TaskAnalysis,
-} from '@layer-ai/types';
+
+// Re-export all types from types package
+export * from './types/index.js';
