+++ packages/sdk/src/resources/keys.ts
@@ -1,5 +1,5 @@
 import type { Layer } from '../client.js';
-import type { ApiKey, CreateKeyRequest, CreateKeyResponse } from '@layer-ai/types';
+import type { ApiKey, CreateKeyRequest, CreateKeyResponse } from '../types/index.js';
 
 export class KeysResource {
   constructor(private client: Layer) {}
